requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosmtplib==2.0.2
python-dotenv==1.0.1
email-validator==2.1.0
pydantic-settings==2.2.1
SQLAlchemy>=2.0.36